pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.57.1",
    "groq>=0.30.0",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.17",
    "langchain-google-genai>=2.1.8",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langchain-tavily>=0.2.7",
    "langgraph>=0.5.3",
    "mcp>=1.11.0",
    "python-dotenv>=1.1.1",
]
